+++ Libft/ft_memset.c
@@ -1,4 +1,6 @@
-void	*ft_memset(void *ptr, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memset(void *ptr, int c, size_t n)
 {
 	n--;
 	while (n > 0)
